move game camera to separate file
